Edit profile/index view so settings are also available for viewing
I edited the profile/index view in a way that the settings are accesable with JS code when pressing the settings button. So the 'settings' view is also ready.
I only have to adjust some things later to make the buttons functional and the language 'selection box' that is now just a p-tag temporarily.